Refactor posts retrieval to use Edge-compatible APIs; remove legacy posts-node module
