--- pages/api/posts-node.ts
@@ -1,7 +1,0 @@
-import type { NextApiRequest, NextApiResponse } from 'next';
-import { getAllPosts } from '../../lib/posts-node';
-
-export default async function handler(req: NextApiRequest, res: NextApiResponse) {
-  const posts = await getAllPosts();
-  res.status(200).json(posts);
-}
